Refactor kafka tests to new directory, with clear separation between local and cloud.
